refactor: move GUI classes into own package
- Classes in the source root was slightly messy
- Also in preparation for further refactoring to abstract Ui related
  logic out of Hazell main class
